requirements.txt
dj-database-url==0.4.2
Django==1.11
django-webpack-loader==0.4.1
djangorestframework==3.6.2
gunicorn==19.7.1
httpie==0.9.9
psycopg2==2.7.1
Pygments==2.2.0
pytz==2017.2
requests==2.13.0
requests-cache==0.4.13
six==1.10.0
wheel==0.24.0
whitenoise==3.3.0
